change $tag_char to single char
changed to single char because the regex that i can currently muster is only able to do that. it's of course less robust because a single '<' is more likely to appear in a title than '<<' but what can you do... changed from '\[\[' because sed's `\[^\[\[` doesn't match '[[' but just a single '[' which is problematic because it appears in a lot of song titles.